+++ internal/controller/cluster_monitoring_controller.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/blang/semver"
+	"github.com/blang/semver/v4"
 	"github.com/pkg/errors"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/types"
@@ -40,18 +40,18 @@ var (
 // ClusterMonitoringReconciler reconciles a Cluster object
 type ClusterMonitoringReconciler struct {
 	// Client is the controller client.
-	client.Client
-	common.ManagementCluster
+	Client            client.Client
+	ManagementCluster common.ManagementCluster
 	// PrometheusAgentService is the service for managing PrometheusAgent resources.
-	prometheusagent.PrometheusAgentService
+	PrometheusAgentService prometheusagent.PrometheusAgentService
 	// AlloyService is the service which manages Alloy monitoring agent configuration.
 	AlloyService alloy.Service
 	// HeartbeatRepository is the repository for managing heartbeats.
-	heartbeat.HeartbeatRepository
+	HeartbeatRepository heartbeat.HeartbeatRepository
 	// MimirService is the service for managing mimir configuration.
-	mimir.MimirService
+	MimirService mimir.MimirService
 	// BundleConfigurationService is the service for configuring the observability bundle.
-	*bundle.BundleConfigurationService
+	BundleConfigurationService *bundle.BundleConfigurationService
 	// MonitoringConfig is the configuration for the monitoring package.
 	MonitoringConfig monitoring.Config
 }
@@ -81,7 +81,6 @@ func SetupClusterMonitoringReconciler(mgr manager.Manager, conf config.Config) e
 	alloyService := alloy.Service{
 		Client:                 managerClient,
 		OrganizationRepository: organizationRepository,
-		PasswordManager:        password.SimpleManager{},
 		ManagementCluster:      conf.ManagementCluster,
 		MonitoringConfig:       conf.Monitoring,
 	}
@@ -180,7 +179,7 @@ func (r *ClusterMonitoringReconciler) Reconcile(ctx context.Context, req ctrl.Re
 	}
 
 	// Handle deletion reconciliation loop.
-	if !cluster.ObjectMeta.DeletionTimestamp.IsZero() {
+	if !cluster.DeletionTimestamp.IsZero() {
 		logger.Info("handling deletion for cluster")
 		return r.reconcileDelete(ctx, cluster)
 	}