+++ pkg/monitoring/alloy/service.go
@@ -10,12 +10,11 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 	"sigs.k8s.io/controller-runtime/pkg/log"
 
-	"github.com/blang/semver"
+	"github.com/blang/semver/v4"
 	"github.com/pkg/errors"
 
 	"github.com/giantswarm/observability-operator/pkg/common"
 	"github.com/giantswarm/observability-operator/pkg/common/organization"
-	"github.com/giantswarm/observability-operator/pkg/common/password"
 	"github.com/giantswarm/observability-operator/pkg/common/tenancy"
 	"github.com/giantswarm/observability-operator/pkg/monitoring"
 )
@@ -28,7 +27,6 @@ const (
 type Service struct {
 	client.Client
 	organization.OrganizationRepository
-	PasswordManager password.Manager
 	common.ManagementCluster
 	MonitoringConfig monitoring.Config
 }
@@ -86,13 +84,13 @@ func (a *Service) ReconcileDelete(ctx context.Context, cluster *clusterv1.Cluste
 	logger.Info("alloy-service - ensuring alloy is removed")
 
 	configmap := ConfigMap(cluster)
-	err := a.Client.Delete(ctx, configmap)
+	err := a.Delete(ctx, configmap)
 	if err != nil && !apierrors.IsNotFound(err) {
 		return errors.WithStack(err)
 	}
 
 	secret := Secret(cluster)
-	err = a.Client.Delete(ctx, secret)
+	err = a.Delete(ctx, secret)
 	if err != nil && !apierrors.IsNotFound(err) {
 		return errors.WithStack(err)
 	}