+++ CHANGELOG.md
@@ -7,9 +7,35 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.26.1] - 2025-04-23
+
+### Fixed
+
+- Fix golangci-lint v2 problems.
+
+## [0.26.0] - 2025-04-23
+
 ### Added
 
-- Add support for loading rules in a tenant in the Mimir Ruler from Workload Clusters.
+- Add validation webhook to validate the alertmanager config before it is send to the alertmanager.
+
+### Fixed
+
+- Ensure support for loading Prometheus Rules in the Mimir Ruler from workload clusters is only enabled for observability-bundle version 1.9.0 and above (extra query matchers have been added in alloy 1.5.0).
+
+## [0.25.0] - 2025-04-22
+
+### Added
+
+- Add support for loading Prometheus Rules in the Mimir Ruler from workload clusters.
+
+### Changed
+
+- Load Prometheus Rules in the Mimir Ruler via Alloy Metrics instead of Alloy Rules on management clusters.
+
+### Removed
+
+- Remove loading of Prometheus Rules for logs into the Loki Ruler via Alloy Rules as it is now managed by Alloy Logs.
 
 ## [0.24.0] - 2025-04-15
 
@@ -421,7 +447,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initialize project and create heartbeat for the installation.
 
-[Unreleased]: https://github.com/giantswarm/observability-operator/compare/v0.24.0...HEAD
+[Unreleased]: https://github.com/giantswarm/observability-operator/compare/v0.26.1...HEAD
+[0.26.1]: https://github.com/giantswarm/observability-operator/compare/v0.26.0...v0.26.1
+[0.26.0]: https://github.com/giantswarm/observability-operator/compare/v0.25.0...v0.26.0
+[0.25.0]: https://github.com/giantswarm/observability-operator/compare/v0.24.0...v0.25.0
 [0.24.0]: https://github.com/giantswarm/observability-operator/compare/v0.23.2...v0.24.0
 [0.23.2]: https://github.com/giantswarm/observability-operator/compare/v0.23.1...v0.23.2
 [0.23.1]: https://github.com/giantswarm/observability-operator/compare/v0.23.0...v0.23.1